Transaction changes
Information from checking in and out are automatically pushed to Transaction database.

Using DateTime.now instead of DateTime.Today to get the time of check in/checkout

Using random number generator to create transaction num.

Removed an earlier addition that made "vacant" the default value for a tempRoom in RoomInfo.